split native.dart
